98 fi

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -5,16 +5,16 @@
  */
 int main(void)
 {
-unsigned long long int a, b, c, i;
+long double a, b, c, i;
 a = 0;
 b = 1;
 c = 1;
 for (i = 1; i <= 98; i++)
 {
 if (i != 98)
-printf("%llu, ", c);
+printf("%Lf, ", c);
 else
-printf("%llu", c);
+printf("%Lf", c);
 a = b;
 b = c;
 c = a + b;
